requirements.txt
scapy
prettytable
python-nmap
numpy
pandas

